#3732 Fixed logic to perform as intended
