+++ backend/audit/utils.py
@@ -23,6 +23,25 @@ class Util:
             # FIXME This should raise an error
             return f"NOT AN ARRAY: {json_array}"
 
+    @staticmethod
+    def remove_extra_fields(general_information_data):
+        """Remove unnecessary fields from general information data."""
+        # Remove unnecessary fields based on auditor_country and auditor_international_address
+        # If auditor country is USA, remove the international address field
+        if general_information_data["auditor_country"] == "USA":
+            general_information_data.pop("auditor_international_address", None)
+        # If auditor country is not USA, remove the USA address fields
+        elif general_information_data["auditor_country"] != "USA":
+            general_information_data.pop("auditor_address_line_1", None)
+            general_information_data.pop("auditor_city", None)
+            general_information_data.pop("auditor_state", None)
+            general_information_data.pop("auditor_zip", None)
+        # Remove unnecessary fields based on audit_period_covered
+        # If audit_period_covered is not "other", remove the audit_period_other_months field
+        if general_information_data["audit_period_covered"] != "other":
+            general_information_data.pop("audit_period_other_months", None)
+        return general_information_data
+
 
 class ExcelExtractionError(Exception):
     def __init__(
